split io into readers and writers
